Fix font issues
