+++ pages/docs/migrate-to-docs-v2.mdx
@@ -21,7 +21,7 @@ import { motion } from "framer-motion";
 
 新的文檔代表著我們將以新的方式和大家見面！如果您有興趣協助我們將文檔變得更完善，那你來對地方了，請往下閱讀如何編輯，想繼續探索機器人功能可以先看看 [快速入門](/docs) 頁面。
 
-<div className="steps-container">
+<div className="steps-container mt-9">
 ### 我的頁面在哪裡??
 
 所有 mdx 檔案都已移至 `/pages`
